migrate the rest of the utilities from __init__.py to utils.py
